refactor: refactor pages and context and routes
