+++ src/hooks/useFetch.ts
@@ -1,4 +1,4 @@
-import { TFormValues } from "../pages/MaterialUiForm";
+import { TFormValues } from "../pages/MaterialUiCadastro";
 
 export const useFetch = async (): Promise<TFormValues> => {
   const data = await fetch("http://localhost:4403/user");
